refactor: PluginAdapter interface for plugins
