+++ CMakeLists.txt
@@ -118,8 +118,6 @@ add_library(
     src/AccessControl.cpp
     src/Client.cpp
     src/Crypto.cpp
-    src/CustomAccessControl.cpp
-    src/CustomLogger.cpp
     src/DataType.cpp
     src/Event.cpp
     src/Logger.cpp
