+++ include/open62541pp/AccessControl.h
@@ -6,6 +6,8 @@
 
 #include "open62541pp/Bitmask.h"
 #include "open62541pp/Span.h"
+#include "open62541pp/detail/open62541/common.h"  // UA_AccessControl
+#include "open62541pp/plugins/PluginAdapter.h"
 #include "open62541pp/types/Builtin.h"
 #include "open62541pp/types/Composed.h"  // UserTokenPolicy, PerformUpdateType
 #include "open62541pp/types/NodeId.h"
@@ -44,18 +46,8 @@ struct Login {
  * @see UA_AccessControl
  * @see https://www.open62541.org/doc/1.3/plugin_accesscontrol.html
  */
-class AccessControlBase {
+class AccessControlBase : public PluginAdapter<UA_AccessControl> {
 public:
-    AccessControlBase() = default;
-
-    virtual ~AccessControlBase() = default;
-
-    AccessControlBase(const AccessControlBase&) = default;
-    AccessControlBase(AccessControlBase&&) noexcept = default;
-
-    AccessControlBase& operator=(const AccessControlBase&) = default;
-    AccessControlBase& operator=(AccessControlBase&&) noexcept = default;
-
     /**
      * Get available user token policies.
      * If the `securityPolicyUri` is empty, the highest available security policy will be used to
@@ -126,6 +118,9 @@ public:
         DateTime endTimestamp,
         bool isDeleteModified
     ) = 0;
+
+    void clear(UA_AccessControl& ac) noexcept override;
+    UA_AccessControl create() override;
 };
 
 /* ----------------------------------- Default access control ----------------------------------- */
