+++ src/ServerConfig.h
@@ -12,9 +12,9 @@
 #include "open62541pp/types/Builtin.h"
 #include "open62541pp/types/Composed.h"  // UserTokenPolicy
 
-#include "CustomAccessControl.h"
 #include "CustomDataTypes.h"
-#include "CustomLogger.h"
+#include "plugins/PluginManager.h"
+#include "plugins/LoggerAdapter.h"
 
 namespace opcua {
 
@@ -27,21 +27,23 @@ public:
         : config_(config) {}
 
     void setLogger(Logger logger) {
-        customLogger_.set(config_.logger, std::move(logger));
+        if (logger) {
+            logger_.assign(LoggerAdapter(std::move(logger)));
+        }
     }
 
     void setCustomDataTypes(std::vector<DataType> dataTypes) {
-        customDataTypes_.set(config_.customDataTypes, std::move(dataTypes));
+        customDataTypes_.assign(std::move(dataTypes));
     }
 
     void setAccessControl(AccessControlBase& accessControl) {
-        customAccessControl_.setAccessControl(config_.accessControl, accessControl);
+        accessControl_.assign(&accessControl);
         setHighestSecurityPolicyForUserTokenTransfer();
         copyUserTokenPoliciesToEndpoints();
     }
 
     void setAccessControl(std::unique_ptr<AccessControlBase> accessControl) {
-        customAccessControl_.setAccessControl(config_.accessControl, std::move(accessControl));
+        accessControl_.assign(std::move(accessControl));
         setHighestSecurityPolicyForUserTokenTransfer();
         copyUserTokenPoliciesToEndpoints();
     }
@@ -98,9 +100,9 @@ private:
     }
 
     UA_ServerConfig& config_;
-    CustomLogger customLogger_;
-    CustomDataTypes customDataTypes_;
-    CustomAccessControl customAccessControl_;
+    CustomDataTypes customDataTypes_{config_.customDataTypes};
+    PluginManager<UA_Logger> logger_{config_.logger};
+    PluginManager<UA_AccessControl> accessControl_{config_.accessControl};
 };
 
 }  // namespace opcua
