+++ src/detail/helper.cpp
@@ -1,6 +1,7 @@
 #include "open62541pp/detail/helper.h"
 
 #include <cstring>
+#include <cstdio>  // vsnprintf
 
 namespace opcua::detail {
 
@@ -29,6 +30,24 @@ UA_String allocNativeString(std::string_view src) {
         s.data = static_cast<UA_Byte*>(UA_EMPTY_ARRAY_SENTINEL);  // NOLINT
     }
     return s;
+}
+
+std::string toString(const char* format, va_list args) {
+    // NOLINTBEGIN
+    va_list argsCopy{};
+    va_copy(argsCopy, args);
+    const int charsToWrite = std::vsnprintf(nullptr, 0, format, argsCopy);
+    va_end(argsCopy);
+    // NOLINTEND
+    if (charsToWrite < 0) {
+        return {};
+    }
+    std::string buffer(charsToWrite, '\0');
+    const int charsWritten = std::vsnprintf(buffer.data(), buffer.size() + 1, format, args);
+    if (charsWritten < 0 || charsWritten > charsToWrite) {
+        return {};
+    }
+    return buffer;
 }
 
 }  // namespace opcua::detail
