+++ tests/AccessControl.cpp
@@ -3,6 +3,7 @@
 #include <doctest/doctest.h>
 
 #include "open62541pp/AccessControl.h"
+#include "open62541pp/Config.h"
 #include "open62541pp/Server.h"
 #include "open62541pp/Session.h"
 #include "open62541pp/types/DataValue.h"
@@ -10,6 +11,7 @@
 #include "helper/stringify.h"
 
 using namespace opcua;
+
 
 TEST_CASE("AccessControlDefault") {
     Server server;
@@ -112,3 +114,209 @@ TEST_CASE("AccessControlDefault") {
         CHECK(ac.allowHistoryDelete(session, {}, {}, {}, {}));
     }
 }
+
+class AccessControlTest : public AccessControlDefault {
+public:
+    using AccessControlDefault::AccessControlDefault;
+
+    Bitmask<AccessLevel> getUserAccessLevel(
+        [[maybe_unused]] Session& session, [[maybe_unused]] const NodeId& nodeId
+    ) override {
+        throw std::runtime_error("This exception should result in most restrictive access");
+    }
+
+    bool allowDeleteNode(
+        [[maybe_unused]] Session& session, [[maybe_unused]] const DeleteNodesItem& item
+    ) noexcept override {
+        return false;
+    }
+
+    bool allowDeleteReference(
+        [[maybe_unused]] Session& session, [[maybe_unused]] const DeleteReferencesItem& item
+    ) noexcept override {
+        return false;
+    }
+};
+
+#if UAPP_OPEN62541_VER_GE(1, 3)
+TEST_CASE("AccessControl plugin adapter") {
+    Server server;
+    AccessControlTest accessControl;
+    UA_AccessControl native = accessControl.create();
+
+    CHECK(native.context != nullptr);
+    CHECK(native.userTokenPoliciesSize == 1);  // anonymous only
+    CHECK(native.userTokenPolicies != nullptr);
+    CHECK(native.userTokenPolicies[0].tokenType == UA_USERTOKENTYPE_ANONYMOUS);
+
+    CHECK(native.activateSession != nullptr);
+    CHECK(
+        native.activateSession(
+            server.handle(),
+            &native,
+            nullptr,  // endpoint description
+            nullptr,  // secure channel remote certificate
+            nullptr,  // session id
+            nullptr,  // user identity token
+            nullptr  // session context
+        ) == UA_STATUSCODE_GOOD
+    );
+
+    CHECK(native.closeSession != nullptr);
+    CHECK_NOTHROW(native.closeSession(
+        server.handle(),
+        &native,
+        nullptr,  // session id
+        nullptr  // session context
+    ));
+
+    CHECK(native.getUserRightsMask != nullptr);
+    CHECK(
+        native.getUserRightsMask(
+            server.handle(),
+            &native,
+            nullptr,  // session id
+            nullptr,  // session context
+            nullptr,  // node id
+            nullptr  // node context
+        ) == 0xFFFFFFFF
+    );
+
+    CHECK(native.getUserAccessLevel != nullptr);
+    CHECK(
+        native.getUserAccessLevel(
+            server.handle(),
+            &native,
+            nullptr,  // session id
+            nullptr,  // session context
+            nullptr,  // node id
+            nullptr  // node context
+        ) == 0x00  // most restrictive access due to exception
+    );
+
+    CHECK(native.getUserExecutable != nullptr);
+    CHECK(
+        native.getUserExecutable(
+            server.handle(),
+            &native,
+            nullptr,  // session id
+            nullptr,  // session context
+            nullptr,  // method id
+            nullptr  // method context
+        ) == true
+    );
+
+    CHECK(native.getUserExecutableOnObject != nullptr);
+    CHECK(
+        native.getUserExecutableOnObject(
+            server.handle(),
+            &native,
+            nullptr,  // session id
+            nullptr,  // session context
+            nullptr,  // method id
+            nullptr,  // method context
+            nullptr,  // object id
+            nullptr  // object context
+        ) == true
+    );
+
+    CHECK(native.allowAddNode != nullptr);
+    CHECK(
+        native.allowAddNode(
+            server.handle(),
+            &native,
+            nullptr,  // session id
+            nullptr,  // session context
+            nullptr  // item
+        ) == true
+    );
+
+    CHECK(native.allowAddReference != nullptr);
+    CHECK(
+        native.allowAddReference(
+            server.handle(),
+            &native,
+            nullptr,  // session id
+            nullptr,  // session context
+            nullptr  // item
+        ) == true
+    );
+
+    CHECK(native.allowDeleteNode != nullptr);
+    CHECK(
+        native.allowDeleteNode(
+            server.handle(),
+            &native,
+            nullptr,  // session id
+            nullptr,  // session context
+            nullptr  // item
+        ) == false
+    );
+
+    CHECK(native.allowDeleteReference != nullptr);
+    CHECK(
+        native.allowDeleteReference(
+            server.handle(),
+            &native,
+            nullptr,  // session id
+            nullptr,  // session context
+            nullptr  // item
+        ) == false
+    );
+
+    CHECK(native.allowBrowseNode != nullptr);
+    CHECK(
+        native.allowBrowseNode(
+            server.handle(),
+            &native,
+            nullptr,  // session id
+            nullptr,  // session context
+            nullptr,  // node id
+            nullptr  // node context
+        ) == true
+    );
+
+#ifdef UA_ENABLE_SUBSCRIPTIONS
+    CHECK(native.allowTransferSubscription != nullptr);
+    CHECK(
+        native.allowTransferSubscription(
+            server.handle(),
+            &native,
+            nullptr,  // old session id
+            nullptr,  // old session context
+            nullptr,  // new session id
+            nullptr  // new session context
+        ) == true
+    );
+#endif
+
+#ifdef UA_ENABLE_HISTORIZING
+    CHECK(native.allowHistoryUpdateUpdateData != nullptr);
+    CHECK(
+        native.allowHistoryUpdateUpdateData(
+            server.handle(),
+            &native,
+            nullptr,  // session id
+            nullptr,  // session context
+            nullptr,  // node id
+            UA_PERFORMUPDATETYPE_INSERT,
+            nullptr  // value
+        ) == true
+    );
+
+    CHECK(native.allowHistoryUpdateDeleteRawModified != nullptr);
+    CHECK(
+        native.allowHistoryUpdateDeleteRawModified(
+            server.handle(),
+            &native,
+            nullptr,  // session id
+            nullptr,  // session context
+            nullptr,  // node id
+            {},  // start timestamp,
+            {},  // stop timestamp,
+            true
+        ) == true
+    );
+#endif
+}
+#endif
